pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Serg1811 <serg181182@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-debug-toolbar = "^4.4.1"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
djoser = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
